% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPopupVideos.R
\name{addPopupVideos}
\alias{addPopupVideos}
\title{Add video popups to leaflet layers.}
\usage{
addPopupVideos(map, video, group, width = NULL, height = NULL)
}
\arguments{
\item{map}{the \code{leaflet} map to add the popups to.}

\item{video}{A character \code{vector} of file path(s) od the video files.}

\item{group}{the map group to which the popups should be added.}

\item{width}{the width of the video(s) in pixels.}

\item{height}{the height of the video(s) in pixels.}
}
\value{
A \code{leaflet} map.
}
\description{
Add video popups to leaflet layers.
}
