% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafpop-package.R
\docType{package}
\name{leafpop-package}
\alias{leafpop-package}
\title{Include Tables, Images and Graphs in Leaflet Popups}
\description{
Include Tables, Images and Graphs in Leaflet Popups
}
\details{
Creates HTML strings to embed tables, images or graphs in popups
of interactive maps created with packages 'leaflet' or 'mapview'. Handles
local paths to images on the file system or remote urls. Handles graphs created
with 'base' graphics, 'lattice' or 'ggplot2' as well as interactive plots created
with 'htmlwidgets'.
}
\author{
Tim Appelhans, Florian Detsch
\emph{Maintainer:} Tim Appelhans \email{tim.appelhans@gmail.com}
}
\keyword{package}
